added locations. swipe refresh working on box, need to update item/locations

+++ app/src/main/java/com/leopardseal/inventorymanagerapp/data/repositories/ImageRepository.kt
@@ -2,12 +2,18 @@ package com.leopardseal.inventorymanagerapp.data.repositories
 
 import android.content.Context
 import android.net.Uri
-import com.leopardseal.inventorymanagerapp.data.UserPreferences
-import com.leopardseal.inventorymanagerapp.data.network.API.ImageAPI
-import kotlinx.coroutines.flow.first
+import android.util.Log
+import com.leopardseal.inventorymanagerapp.data.network.api.ImageAPI
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.launch
 import okhttp3.MediaType.Companion.toMediaTypeOrNull
 import okhttp3.RequestBody
+import okhttp3.RequestBody.Companion.asRequestBody
 import okhttp3.RequestBody.Companion.toRequestBody
+import okio.BufferedSink
+import java.io.File
 import javax.inject.Inject
 import javax.inject.Singleton
 
@@ -16,21 +22,21 @@ class ImageRepository @Inject constructor(
     private val imageApi : ImageAPI
 ): BaseRepository() {
 
-    suspend fun uploadImage(url : String, image : Uri, context: Context) = safeApiCall {
-        imageApi.uploadImageToBlob(
+    suspend fun uploadImage(url : String, imageFile: File) = safeApiCall {
+//        delay(100)
+        val byteArray = imageFile.readBytes()
+        val requestBody = byteArray.toRequestBody("image/jpeg".toMediaTypeOrNull())
+
+
+        val response = imageApi.uploadImageToBlob(
             sasUrl = url,
-            image = uriToRequestBody(context, image)
+            image = requestBody
         )
-    }
-    fun uriToRequestBody(context: Context, uri: Uri): RequestBody {
-        val inputStream = context.contentResolver.openInputStream(uri)
-        val byteArray = inputStream?.readBytes() ?: byteArrayOf()
-        context.contentResolver.delete(uri, null, null)
-        return byteArray.toRequestBody("image/jpeg".toMediaTypeOrNull())
+        if(response.isSuccessful){
+            Log.d("Upload", "success")
+            imageFile.delete()
+        }
+        response
     }
 
-    fun fileToRequestBody(imageFile : File) : RequestBody{
-        requestBody = imageFile.asRequestBody("image/jpeg".toMediaTypeOrNull())
-        imageFile.delete()
-    }
 }
